rafactor: mission8 코드 리펙토링
