+++ src/services/notificationService.ts
@@ -3,19 +3,19 @@ import { NotificationData } from '../typings/notification';
 import UnauthorizedError from '../lib/errors/UnauthorizedError';
 
 export const notificationService = {
-  markAsRead: (id: number) => {
-    return notificationRepository.markAsRead(id);
+  async markAsRead(id: number) {
+    return await notificationRepository.markAsRead(id);
   },
 
-  markAllAsRead: (user: { id: number } | undefined) => {
+  async markAllAsRead(user: { id: number } | undefined) {
     if (!user) {
       throw new UnauthorizedError('Unauthorized');
     }
 
-    return notificationRepository.markAllAsRead(user.id);
+    return await notificationRepository.markAllAsRead(user.id);
   },
 
   async createNotification(input: NotificationData) {
-    return notificationRepository.create(input);
+    return await notificationRepository.create(input);
   },
 };
